--- src/components/AccordionUI/index.js
@@ -1,1 +1,0 @@
-export { default as AccordionUI } from './AccordionUI';
