+++ src/components/ButtonGroupUI/index.js
@@ -1,0 +1,1 @@
+export { default as ButtonGroupUI } from './ButtonGroupUI';
