+++ src/components/FrontModal/index.js
@@ -1,0 +1,1 @@
+export { default as FrontModal } from './FrontModal';
