+++ src/components/index.js
@@ -1,5 +1,4 @@
 export * from './About';
-export * from './AccordionUI';
 export * from './BookList';
 export * from './BookSelect';
 export * from './Card';
@@ -9,6 +8,7 @@ export * from './ChapterSelect';
 export * from './ContextMenu';
 export * from './DialogUI';
 export * from './Intro';
+export * from './FrontModal';
 export * from './OBSVerses';
 export * from './SelectModeBible';
 export * from './SelectResourcesLanguages';
